requirements.txt
gradio
python-dotenv
pandas
langchain
langchain-community
langchain-core
langchain-google-genai
